✅Migrate match package json test to snaps
